Motor ints to floats

+++ snn_motor_controller/msgs/controlOutputMotorCommands.h
@@ -1,6 +1,6 @@
 struct __attribute__((__packed__)) serial_control_out {
     //Motor commands
-    int16_t motor_1, motor_2, motor_3, motor_4;
+    float motor_1, motor_2, motor_3, motor_4;
     //CHECKSUM
     uint8_t checksum_out;
 };
